requirements.txt
pandas==1.4.4
Pillow==9.2.0
Pillow==9.5.0
plotly==5.9.0
streamlit==1.23.1
